app/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Kolin"]

[tool.poetry.dependencies]
python = "3.10.4"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
SQLAlchemy = "^1.4.36"
httpx = "^0.22.0"
pydantic = "^1.9.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^10.0.2"}
aiosqlite = "^0.17.0"
asyncpg = "^0.25.0"
pip = "^22.0.4"
install = "^1.3.5"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
alembic = "^1.7.7"
bpython = "^0.22.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
